feat: add mediaHeaders
