+++ src/components/Activity/components/ActivityAll.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import useSWR from "swr";
 import { fuegoChartTop6 } from "../config/fuego";
-import type { FilterData } from "../../../types/typesFilters";
 import useIsBreakpoint from "../../../utils/useIsBreakpoint";
 import {
   VizGenre,
@@ -9,6 +8,7 @@ import {
   VizRating,
   VizReleased,
 } from "./components";
+import type { FilterData } from "../../Filters/config";
 
 export default function ActivityAll({
   uid,
