--- src/components/Content/components/ContentFilters/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ContentFilters";
