+++ src/components/Content/components/ContentToolbar/ContentToolbar.tsx
@@ -15,17 +15,17 @@ import React from "react";
 import { useMDDispatch } from "../../../../config/store";
 import useIsBreakpoint from "../../../../utils/useIsBreakpoint";
 import MdLogo from "../../../md/MdLogo";
-import Filters from "../ContentFilters/ContentFilters";
+// import Filters from "../../../Filters/components/ContentFilters";
 
 function ContentToolbar({ onOpen }: { onOpen: () => void }): JSX.Element {
   const isMd = useIsBreakpoint("md");
-  const router = useRouter();
+  // const router = useRouter();
   const { colorMode, toggleColorMode } = useColorMode();
-  const {
-    isOpen: menuIsOpen,
-    onClose: menuOnClose,
-    onOpen: menuOnOpen,
-  } = useDisclosure();
+  // const {
+  //   isOpen: menuIsOpen,
+  //   onClose: menuOnClose,
+  //   onOpen: menuOnOpen,
+  // } = useDisclosure();
   const dispatch = useMDDispatch();
   return (
     <>
@@ -57,7 +57,7 @@ function ContentToolbar({ onOpen }: { onOpen: () => void }): JSX.Element {
             </Flex>
             <Flex maxW="720px" align="center">
               <HStack spacing="2">
-                {router.pathname === "/home" && (
+                {/* {router.pathname === "/home" && (
                   <>
                     {isMd ? (
                       <Button
@@ -79,7 +79,7 @@ function ContentToolbar({ onOpen }: { onOpen: () => void }): JSX.Element {
                       />
                     )}
                   </>
-                )}
+                )} */}
                 <IconButton
                   onClick={toggleColorMode}
                   aria-label="Theme Switcher"
@@ -107,7 +107,7 @@ function ContentToolbar({ onOpen }: { onOpen: () => void }): JSX.Element {
           </Flex>
         </Container>
       </Box>
-      <Filters isOpen={menuIsOpen} onClose={menuOnClose} />
+      {/* <Filters isOpen={menuIsOpen} onClose={menuOnClose} /> */}
     </>
   );
 }
