+++ src/components/Content/components/index.tsx
@@ -1,4 +1,3 @@
 export { default as ContentDrawer } from "./ContentDrawer";
-export { default as ContentFilters } from "./ContentFilters";
 export { default as ContentSidebar } from "./ContentSidebar";
 export { default as ContentToolbar } from "./ContentToolbar";
