+++ src/components/Filters/components/FiltersContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FiltersContent";
