+++ src/components/Filters/components/index.ts
@@ -1,0 +1,1 @@
+export { default as FiltersContent } from "./FiltersContent";
