+++ src/components/Filters/config/index.ts
@@ -1,0 +1,2 @@
+export * from "./fuego";
+export * from "./types";
