+++ src/components/Home/components/HomeHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HomeHeader";
