+++ src/components/Home/components/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeHeader } from "./HomeHeader";
