+++ src/components/Home/config/index.ts
@@ -1,1 +1,2 @@
 export { fuegoDiaryGet } from "./fuego";
+export * from "./types";
