+++ src/components/Media/MediaSpotify.tsx
@@ -71,11 +71,7 @@ export default function MediaSpotify({
                 text={dayjs(diaryDate).format("MMM D, YYYY")}
               />
             )}
-            {rating ? (
-              <MediaRating rating={rating} />
-            ) : (
-              <MediaInfoText title="Rating" text="No Rating" />
-            )}
+            {rating && rating > 0 ? <MediaRating rating={rating} /> : null}
             {release_date && (
               <MediaInfoText
                 title="Released"
