+++ src/components/Memories/components/MemoriesHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MemoriesHeader";
