+++ src/components/Memories/components/index.ts
@@ -1,0 +1,1 @@
+export { default as MemoriesHeader } from "./MemoriesHeader";
