+++ src/components/Memories/config/index.ts
@@ -1,1 +1,2 @@
 export { fuegoMemoriesGet } from "./fuego";
+export type { SortType } from "./types";
