+++ src/components/Memories/config/types.ts
@@ -1,0 +1,1 @@
+export type SortType = "addedDate" | "rating";
