+++ src/config/store.ts
@@ -1,5 +1,8 @@
 import { createContext, useContext } from "react";
-import type { FilterBookmarkState, FilterState } from "../types/typesFilters";
+import type {
+  FilterState,
+  FilterBookmarkState,
+} from "../components/Filters/config";
 import type { MDbMovie, MDbTV } from "../types/typesMDb";
 import type { MediaDiaryWithId, MediaSelected } from "../types/typesMedia";
 import type { SpotifyAlbum, SpotifyArtist } from "../types/typesSpotify";
