+++ src/fuego/index.ts
@@ -1,6 +1,5 @@
 export * from "./fuego";
 export * from "./fuegoBookmarks";
-export * from "./fuegoFilterActions";
 export * from "./fuegoMDActions";
 export * from "./fuegoProvider";
 export * from "./useFuegoAuth";
